Code restructure
