+++ src/reader/utils/mod.rs
@@ -3,6 +3,7 @@ use geometry::Point;
 use super::binarize::BinaryImage;
 
 pub mod accumulate;
+pub mod frame;
 pub mod geometry;
 pub mod homography;
